chore: move mock slow data storage to the mock package
