+++ pkg/server/test/list_objects.go
@@ -12,6 +12,7 @@ import (
 	"github.com/openfga/openfga/pkg/logger"
 	"github.com/openfga/openfga/pkg/server/commands"
 	"github.com/openfga/openfga/pkg/storage"
+	"github.com/openfga/openfga/pkg/storage/mocks"
 	"github.com/openfga/openfga/pkg/tuple"
 	"github.com/openfga/openfga/pkg/typesystem"
 	"github.com/stretchr/testify/require"
@@ -27,119 +28,6 @@ type mockStreamServer struct {
 func (x *mockStreamServer) Send(m *openfgapb.StreamedListObjectsResponse) error {
 	x.channel <- m.Object
 	return nil
-}
-
-// mockSlowDataStorage is a proxy to the actual ds except the Reads are slow time by the sleepTime
-// This allows simulating list objection condition that times out
-type mockSlowDataStorage struct {
-	sleepTime time.Duration
-	ds        storage.OpenFGADatastore
-}
-
-func NewMockSlowDataStorage(ds storage.OpenFGADatastore, sleepTime time.Duration) storage.OpenFGADatastore {
-	return &mockSlowDataStorage{
-		sleepTime: sleepTime,
-		ds:        ds,
-	}
-}
-
-func (m *mockSlowDataStorage) Close() {}
-
-func (m *mockSlowDataStorage) ListObjectsByType(ctx context.Context, store string, objectType string) (storage.ObjectIterator, error) {
-	return m.ds.ListObjectsByType(ctx, store, objectType)
-}
-
-func (m *mockSlowDataStorage) Read(ctx context.Context, store string, key *openfgapb.TupleKey) (storage.TupleIterator, error) {
-	time.Sleep(m.sleepTime)
-	return m.ds.Read(ctx, store, key)
-}
-
-func (m *mockSlowDataStorage) ReadPage(ctx context.Context, store string, key *openfgapb.TupleKey, paginationOptions storage.PaginationOptions) ([]*openfgapb.Tuple, []byte, error) {
-	time.Sleep(m.sleepTime)
-	return m.ds.ReadPage(ctx, store, key, paginationOptions)
-}
-
-func (m *mockSlowDataStorage) ReadChanges(ctx context.Context, store, objectType string, paginationOptions storage.PaginationOptions, horizonOffset time.Duration) ([]*openfgapb.TupleChange, []byte, error) {
-	return m.ds.ReadChanges(ctx, store, objectType, paginationOptions, horizonOffset)
-}
-
-func (m *mockSlowDataStorage) Write(ctx context.Context, store string, deletes storage.Deletes, writes storage.Writes) error {
-	return m.ds.Write(ctx, store, deletes, writes)
-}
-
-func (m *mockSlowDataStorage) ReadUserTuple(ctx context.Context, store string, key *openfgapb.TupleKey) (*openfgapb.Tuple, error) {
-	time.Sleep(m.sleepTime)
-	return m.ds.ReadUserTuple(ctx, store, key)
-}
-
-func (m *mockSlowDataStorage) ReadUsersetTuples(ctx context.Context, store string, filter storage.ReadUsersetTuplesFilter) (storage.TupleIterator, error) {
-	time.Sleep(m.sleepTime)
-	return m.ds.ReadUsersetTuples(ctx, store, filter)
-}
-
-func (m *mockSlowDataStorage) ReadStartingWithUser(
-	ctx context.Context,
-	store string,
-	filter storage.ReadStartingWithUserFilter,
-) (storage.TupleIterator, error) {
-	time.Sleep(m.sleepTime)
-	return m.ds.ReadStartingWithUser(ctx, store, filter)
-}
-
-func (m *mockSlowDataStorage) ReadAuthorizationModel(ctx context.Context, store string, id string) (*openfgapb.AuthorizationModel, error) {
-	return m.ds.ReadAuthorizationModel(ctx, store, id)
-}
-
-func (m *mockSlowDataStorage) ReadAuthorizationModels(ctx context.Context, store string, options storage.PaginationOptions) ([]*openfgapb.AuthorizationModel, []byte, error) {
-	return m.ds.ReadAuthorizationModels(ctx, store, options)
-}
-
-func (m *mockSlowDataStorage) FindLatestAuthorizationModelID(ctx context.Context, store string) (string, error) {
-	return m.ds.FindLatestAuthorizationModelID(ctx, store)
-}
-
-func (m *mockSlowDataStorage) ReadTypeDefinition(ctx context.Context, store, id, objectType string) (*openfgapb.TypeDefinition, error) {
-	return m.ds.ReadTypeDefinition(ctx, store, id, objectType)
-}
-
-func (m *mockSlowDataStorage) WriteAuthorizationModel(ctx context.Context, store string, model *openfgapb.AuthorizationModel) error {
-	return m.ds.WriteAuthorizationModel(ctx, store, model)
-}
-
-func (m *mockSlowDataStorage) CreateStore(ctx context.Context, newStore *openfgapb.Store) (*openfgapb.Store, error) {
-	return m.ds.CreateStore(ctx, newStore)
-}
-
-func (m *mockSlowDataStorage) DeleteStore(ctx context.Context, id string) error {
-	return m.ds.DeleteStore(ctx, id)
-}
-
-func (m *mockSlowDataStorage) WriteAssertions(ctx context.Context, store, modelID string, assertions []*openfgapb.Assertion) error {
-	return m.ds.WriteAssertions(ctx, store, modelID, assertions)
-}
-
-func (m *mockSlowDataStorage) ReadAssertions(ctx context.Context, store, modelID string) ([]*openfgapb.Assertion, error) {
-	return m.ds.ReadAssertions(ctx, store, modelID)
-}
-
-func (m *mockSlowDataStorage) MaxTuplesPerWrite() int {
-	return m.ds.MaxTuplesPerWrite()
-}
-
-func (m *mockSlowDataStorage) MaxTypesPerAuthorizationModel() int {
-	return m.ds.MaxTypesPerAuthorizationModel()
-}
-
-func (m *mockSlowDataStorage) GetStore(ctx context.Context, storeID string) (*openfgapb.Store, error) {
-	return m.ds.GetStore(ctx, storeID)
-}
-
-func (m *mockSlowDataStorage) ListStores(ctx context.Context, paginationOptions storage.PaginationOptions) ([]*openfgapb.Store, []byte, error) {
-	return m.ds.ListStores(ctx, paginationOptions)
-}
-
-func (m *mockSlowDataStorage) IsReady(ctx context.Context) (bool, error) {
-	return m.ds.IsReady(ctx)
 }
 
 type listObjectsTestCase struct {
@@ -276,7 +164,7 @@ func TestListObjectsRespectsMaxResults(t *testing.T, ds storage.OpenFGADatastore
 		{
 			// should not return any useful data as it times out
 			// however, there should not be any error
-			name:   "connected objects timeout",
+			name:   "list_objects_timeout",
 			schema: typesystem.SchemaVersion1_1,
 			model: `
 			type user
@@ -331,7 +219,7 @@ func TestListObjectsRespectsMaxResults(t *testing.T, ds storage.OpenFGADatastore
 
 			datastore := ds
 			if test.dsSlowTime > 0 {
-				datastore = NewMockSlowDataStorage(ds, test.dsSlowTime)
+				datastore = mocks.NewMockSlowDataStorage(ds, test.dsSlowTime)
 			}
 
 			listObjectsQuery := &commands.ListObjectsQuery{
